Added healthbars

+++ final/unityproject/Assets/Scripts/Constants.cs
@@ -1,10 +1,10 @@
 ﻿public static class Constants
 {
     // HEALTH
-    public static int PLAYER_MAX_BASE_HEALTH = 100;
-    public static int BASE_MAX_BASE_HEALTH = 10000;
-    public static int TOWER_MAX_BASE_HEALTH = 5000;
-    public static int MINION_MAX_BASE_HEALTH = 50;
+    public static float PLAYER_MAX_BASE_HEALTH = 100.0f;
+    public static float BASE_MAX_BASE_HEALTH = 10000.0f;
+    public static float TOWER_MAX_BASE_HEALTH = 5000.0f;
+    public static float MINION_MAX_BASE_HEALTH = 50.0f;
     // SPEED
     public static float PLAYER_BASE_VELOCITY = 500.0f;
     public static float PLAYER_BASE_ROTATION_SPEED = 300.0f;
